Update 0.31
Added Object Pooler, Item Manager, Crosshair, and Weapon System. All of which feed into the Controller Class and the balistics class.

The object pooler allows a set amount of objects to be instantiated and runs though said objects and adjusts their values according to new data. This allows for expensive calls like Instantiating and Destroying gameobjects to be set to a minimum. In it's current state, the object pooler is used to display bullet lines which act as player feedback.

The item manager allows for the creation of item types such as different weapons to be created. These are stored in a dictionary and can be called when needed.

The crosshair class is very simple, as it is designed to adjust the spread of the crosshair according to the absolute value of the players movement (ie. players current velocity, and movement of the mouse)

Made adjustments to the Controller class to feed into the other classes and act as a central hub or "driving class" for the systems in place. Now, the player can move, shoot, & look. Not crazy, but must of this update was spent making sure there are foundational systems in place for the rest of the project.